Refactor codebase into modular package
