+++ modchecker/__init__.py
@@ -1,0 +1,4 @@
+from .models import ModInfo, VersionCheckResult
+from .compatibility import find_common_version
+
+__all__ = ["ModInfo", "VersionCheckResult", "find_common_version"]
